rgw_sal_motr: implement watch-notify mechanism using Motr FDMI
The watch-notify can be used to watch (monitor) any update on objects.
A notification is sent (to all rgw instances) by a notifier if an object
is changed, such as object size, object content etc. The notification sent
contains marker for a watcher to pick it up and it also includes required
information for the watcher to act accordingly.

This POC shows how to implement the watch-notify mechanism using FDMI as
messge passing, filtering and notifying. We also show its usage in object
metadata cache uses where it is used to inform object changes and takes
action on cache items.

Signed-off-by: Sining Wu <sining.wu@seagate.com>